requirements.txt
absl-py==0.14.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==11.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
six==1.15.0
tb-nightly==2.7.0a20211001
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tf-estimator-nightly==2.7.0.dev2021092408
tifffile==2021.8.30
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
